Refactor code to move file functionality into separate files.
